Inital commit; add basic accounts view
This is the inital commit for the repo. It is organized into three
different projects, for the Model, the ViewModels, and the Views.

Created a basic accounts list view and gave it a simple view model. At the
moment fake data is being used to have something to look at.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>