Begin work on defining the model
Installed the Entity Framework Core packages.

Started working on the model. I ran into a couple of potential issues:

1) Will decimal work as a type? I remember having some kind of trouble
with it when I was doing my first UWP app, but I don't remember the
details. I should do some experiments before I go any further.

2) The definition of Lot Assignment makes reference to two different
Transactions. I am not sure how to do this in EFCore. The way to refer
to a single Transaction would be with a property called "TransactionId"
but obviously I can't do this twice. I might need to use the attributes
(or whatever the [BlahBlahBlah] things are called).

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>