Create the first migration
Since this is the first migration, and I am doing a more complicate model
than I have ever done before, I am putting it on a branch so that it will
be easy to unroll if necessary.

I wonder if I can put in some unit tests to make sure the model works as
expected. This is another thing that I have never done before.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>