Replay migration due to model changes
I changed the model in order to add more relationships. At least one of these
changes (adding Security to Price) resulted in a foreign key being added.
SQLite cannot handle this in a migration without the developer adding a bunch
of code, so I just did what I have done before: I removed all of the migrations
and started over. This works because the new first migration (Migration_1B)
can do it all at the same time (it is the adding to an existing schema that
is the problem).

I also added a couple more unit tests. So far the model is working exactly as
I had expected it to.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>