Add use of real database
This is the first time the code has tried to access a real database file.
The required references were added to the projects.
The account list view gets the open accounts from the database and adds
them to the list of temp accounts that are hard-coded.

Because the app does not support creating a new account, I just copied
a database file from the unit test. It has a bunch of accounts in it, which
causes the list to overflow the view. No scrollbar appears, which is a bug
that I need to fix soon.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>