Update to use ModelService for basic account info
The ModelService now has methods for account access and modification. The
VMs use these methods instead of accessing the database directly. The ability
to change the Name and Institution of the selected Account was added; this
needs to be changed in the future for better UX.

Packages were updated in response to updating Visual Studio to 15.9.3.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>