Add cash balance calculation to transactions list
This implementation uses ListcollectionView, which has something to do with
how WPF sees the data, as opposed to the underlying data itself. It feels
weird doing it this way, but on the other hand 'cash balance' is not an
attribute of the model, so there should be no need for the model itself to
change, even if just to sort the transactions in an account.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>