Change Security implementation and remove migration to prepare for new Model
The way Securities are handled has been changed. Instead of the user classes
having a direct reference to a Security object, they have only the ID of the
Security object. This is done so that EF doesn't think there is a relationship
between the two classes. This is what has been causing me problems; whenever
Security was changed in Transaction, EF thought that there was a 1:1
relationship to Security, and so tried to create a new Security object with an
ID that was already in use. An alternative solution would be to explicitly make
a 1:N relationship, in which the Security would have a collection containing
every Transaction (and Price, and Position) that uses the Security. This seems
like overkill, and it also doesn't match the way the objects relate. The
Security table is really just a look-up dictionary, so it seems wrong to model
it as a formal relationship.

The existing migration was deleted. A new one will be built for the next
commit.

The build errors were fixed, but no attempt was made to make the code correct.
That will come in subsequent commits.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>