Fix Issue#13 - Display blank Security field for Deposit/Withdrawal
The root of the problem was that the combo list did not have an entry for
"", which is what Deposit and Withdrawal now use. Because of this, the View
stayed with what was already there.

The fix was to add a "" entry at the front of the combo list. This seems
like a bit of a hack to me, but it gets the job done and it doesn't look
too bad so I am going to keep it until I can think of something better.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>