Fix Issue#17: Enforce one price per date in AddPrice
The database is only going to store one price per date per security.
The new price will replace the old price, unless the old price is a closing
price and the new price is a non-closing price.
The prices in the transactions do not change. This is just for historical use.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>