Implement a global positions list
Two new classes are added:
- GlobalPositions_VM
- GlobalPositions_View.

Add new ViewModel - GlobalPositions_VM

GlobalPositions_VM aggregates the positions in all of the open accounts
into one list. The expectation is that the user will likely want to see
everything that they own in a single list, as opposed to clicking through
multiple accounts, which is more work and isn't very useful when the same
security is owned in multiple accounts.

Unit tests for the ViewModel are included.

Add new View - GlobalPositions_View

Created a new View called GlobalPositions_View to connect to
GlobalPositions_VM. The positions are shown in a bare-bones GridView
at the moment. The view is displayed in a new tab on the main window.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>