Add useful info to the Global Positions tab
"Useful" in this context means things the user wants to know; specifically,
the Symbol and the Security Name. To do this I created a new View Model
to represent each entry, GlobalPosition_VM. This includes the Position
object from the Model, as well as Symbol and SecurityName.

The GlobalPositions_View still shows the database IDs, and the grid is not
very pretty, but it is good enough for now.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>