Implement ModelService.GetAccountValue()
This method returns the current value of the account, based on the amount of
cash and the current value of all positions. Unit tests have been added for
this method.
AddTransaction() was updated to track the Cash in the account, which it hadn't
been doing before. With this correction, Account_VM.CalculateCashBalance() no
longer needs to manually change the value of TheAccount.Cash.
Account_VM now has a Value property, which the account list shows in the main
window.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>