Fix Issue23 - Update Deposit doesn't work
The root of the problem was that the Transaction objects had different
versions of the Account object. This meant that when the deposit amount
was being backed out, the value of Account.Cash wasn't consistent.

The solution appears to be to always use the objects that are already
in memory. Fetching a new object from the database might result in a
synchronization issue. This was present in the app code as well as in
some of the unit tests. All of these occurances have been fixed.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>