Handle the cases where a Position goes away
The existing code wasn't handling the cases where a Position was established
and then closed by subsequent transactions (the standard way is to Buy and
then Sell, but it is also valid to Sell and then Buy). A Position object
remained with a Quantity of 0. This caused some problems elsewhere in the code.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>