WIP First attempt at LotAssignment management
This code compiles but some unit tests fail!

It looks like I am running into some of the EF issues I have seen before.
It might be necessary to take a break and work on some more samples in order
to figure out how I should be doing things.

I have also beeing thinking that the schema might need to change in order to
handle lots better. There could be a Lot class that would be similar to the
Position class. This mean more work when processing a transaction, but there
will be less work by the methods that need to know the lots and currently
construct them by going through all Buy and Sell transactions.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>